REFACTOR(type): infrastructure 패키지명 오타 수정
resolved #6
