+++ src/main/java/kr/jojoba/finance/wallet/application/WalletService.java
@@ -1,6 +1,6 @@
 package kr.jojoba.finance.wallet.application;
 
-import kr.jojoba.finance.wallet.infrastrucrture.WalletRepository;
+import kr.jojoba.finance.wallet.infrastructure.WalletRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
