--- src/main/java/kr/jojoba/finance/wallet/infrastrucrture/WalletCustomRepository.java
@@ -1,4 +1,0 @@
-package kr.jojoba.finance.wallet.infrastrucrture;
-
-public interface WalletCustomRepository {
-}
