+++ src/main/java/kr/jojoba/finance/wallet/infrastructure/WalletCustomRepository.java
@@ -1,0 +1,4 @@
+package kr.jojoba.finance.wallet.infrastructure;
+
+public interface WalletCustomRepository {
+}
