+++ src/main/java/kr/jojoba/finance/wallet/infrastructure/WalletRepository.java
@@ -1,4 +1,4 @@
-package kr.jojoba.finance.wallet.infrastrucrture;
+package kr.jojoba.finance.wallet.infrastructure;
 
 import kr.jojoba.finance.wallet.domain.Wallet;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,5 +6,5 @@ import org.springframework.data.repository.history.RevisionRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WalletRepository extends JpaRepository<Wallet, String>, RevisionRepository<Wallet, String, Integer>, WalletCustomRepository {
+public interface WalletRepository extends JpaRepository<Wallet, String>, RevisionRepository<Wallet, String, Integer>, kr.jojoba.finance.wallet.infrastructure.WalletCustomRepository {
 }
